Decouple DetectResource from Params.
Move the DetectResource function out of Params and into its own source
file.  It will become more complex and will have its own unit tests
in the future with the "real" algorithm.
