+++ src/Makefile.am
@@ -31,6 +31,7 @@ libtaurion_la_SOURCES = \
   pending.cpp \
   prospecting.cpp \
   protoutils.cpp \
+  resourcedist.cpp \
   spawn.cpp
 libtaurionheaders = \
   amount.hpp \
@@ -48,6 +49,7 @@ libtaurionheaders = \
   pending.hpp \
   prospecting.hpp \
   protoutils.hpp \
+  resourcedist.hpp \
   spawn.hpp
 
 tauriond_CXXFLAGS = \
