+++ src/params.hpp
@@ -23,11 +23,9 @@
 
 #include "hexagonal/coord.hpp"
 #include "database/faction.hpp"
-#include "database/inventory.hpp"
 #include "proto/character.pb.h"
 
 #include <xayagame/gamelogic.hpp>
-#include <xayautil/random.hpp>
 
 #include <string>
 #include <vector>
@@ -130,13 +128,6 @@ public:
    * demo competition.
    */
   const std::vector<PrizeData>& ProspectingPrizes () const;
-
-  /**
-   * Determines the type and initial amount of resource mine-able that should
-   * be found by prospecting in the given coordinate.
-   */
-  void DetectResource (const HexCoord& pos, xaya::Random& rnd,
-                       std::string& type, Inventory::QuantityT& amount) const;
 
   /**
    * Returns the spawn centre and radius for the given faction.
